lib/core: FileReader checks the end of input file; split Serilizer and Deserilizer
